class validator and user controller

+++ controllers/userController.ts
@@ -1,4 +1,6 @@
 import User from '../models/user';
+import UserValidator from '../classValidators/userClassValidator';
+import { validateOrRejectExample } from '../classValidators/validation';
 
 // const User = require('../models/user');
 
@@ -7,7 +9,27 @@ const getUsers = async (req, res) => {
         const users = await User.find()
         res.status(200).json({ users })
     } catch (error) {
-        throw error
+        console.log("myerror");
+        console.log(error);
+        res.status(403).json("Fordidden");
+    }
+}
+
+const getUser = async (req, res) => {
+    try {
+        const {
+            params: { id },
+        } = req;
+        const user = await User.findOne({_id: id})
+        if(!user) {
+            res.status(404).json("Id doesn't exist");
+            return;
+        }
+        res.status(200).json({ user })
+    } catch (error) {
+        console.log("myerror");
+        console.log(error);
+        res.status(403).json("Fordidden");
     }
 }
 
@@ -23,15 +45,18 @@ const addUser = async (req, res) => {
             email: body.email
         }) 
 
+        const userValidator = new UserValidator(user);
+        await validateOrRejectExample(userValidator);
+
         const newUser = await user.save()
         const allUsers = await User.find()
 
-        console.log("newUser");
-        console.log(newUser);
-
-        res.status(201).json({ message: 'User added', user: newUser, users: allUsers })
+        // res.status(201).json({ message: 'User added', user: newUser, users: allUsers })
+        res.status(201).json({ user: newUser })
     } catch (error) {
-        throw error
+        console.log("myerror");
+        console.log(error);
+        res.status(403).json("Fordidden");
     }
 }
 
@@ -41,38 +66,75 @@ const updateUser = async (req, res) => {
             params: { id },
             body,
         } = req
+
+        const userValidator = new UserValidator(body);
+        await validateOrRejectExample(userValidator);
+
+        const user = await User.findOne({_id: id})
+        if(!user) {
+            res.status(404).json("Id doesn't exist");
+            return;
+        }
+
         const updateUser = await User.findByIdAndUpdate(
             { _id: id },
             body
         )
-        const allUsers = await User.find()
-        res.status(200).json({
-            message: 'User updated',
-            user: updateUser,
-            users: allUsers,
-        })
+        //const allUsers = await User.find()
+
+        // res.status(200).json({
+        //     message: 'User updated',
+        //     user: updateUser,
+        //     users: allUsers,
+        // })
+        res.status(200).json({user: updateUser})
     } catch (error) {
-        throw error
+        console.log("myerror");
+        console.log(error);
+        res.status(403).json("Fordidden");
     }
 }
 
 const deleteUser = async (req, res) => {
     try {
+        const {params: { id }} = req;
+        const user = await User.findOne({_id: id})
+        if(!user) {
+            res.status(404).json("Id doesn't exist");
+            return;
+        }
+
         const deletedUser = await User.findByIdAndRemove(
-            req.params.id
+            // req.params.id
+            id
         )
-        const allUsers = await User.find()
-        res.status(200).json({
-            message: 'User deleted',
-            user: deletedUser,
-            users:allUsers,
-        })
+
+        // const allUsers = await User.find()
+        // res.status(200).json({
+        //     message: 'User deleted',
+        //     user: deletedUser,
+        //     users:allUsers,
+        // })
+        res.status(200).json({ user: deletedUser });
     } catch (error) {
-        throw error
+        console.log("myerror");
+        console.log(error);
+        res.status(403).json("Fordidden");
+        // throw error
     }
 }
 
-export { getUsers, addUser, updateUser, deleteUser }
+// async function validateOrRejectExample(input) {
+//     try {
+//       await validateOrReject(input);
+//       return 'OK'
+//     } catch (errors) {
+//     //   console.log('Caught promise rejection (validation failed). Errors: ', errors);
+//       throw errors;
+//     }
+//   }
+
+export { getUsers, getUser, addUser, updateUser, deleteUser }
 
 // module.exports.getUsers = getUsers;
 // module.exports.addUser = addUser;
